RAC4861: SprintTag missed logs of first 9 times test
