+++ jobs/ContinuousTest/Jenkinsfile
@@ -52,7 +52,14 @@ node{
         }
         stage("Test"){
             // Start to run test
-            load("build-config/jobs/function_test/function_test.groovy")
+            def TESTS = "${env.TESTS}"
+            def repo_dir = pwd() + "/build-config"
+            def function_test = load("build-config/jobs/FunctionTest/FunctionTest.groovy")
+            try{
+                function_test.run(TESTS, env.stash_manifest_name, env.stash_manifest_path, repo_dir)
+            } finally{
+                function_test.archiveArtifactsToTarget("FunctionTest")
+            }
         }
     } finally{
         def shareMethod = load("build-config/jobs/ShareMethod.groovy")
