+++ jobs/MasterCI/MasterCI
@@ -14,17 +14,27 @@ node{
         stash name: "masterci_manifest", includes: "${manifest_name}"
         env.stash_manifest_name = "masterci_manifest"
         env.stash_manifest_path = "${manifest_name}"
-       
-        stage("Unit Test"){
-            // Create an instance of UnitTest/UnitTest.groovy
-            def unit_test = load("jobs/UnitTest/UnitTest.groovy")
-            // Call the function runTest to run unit test
-            unit_test.runTest("masterci_manifest", manifest_name, pwd())
-        }
-        stage("Function Test"){
-            load("jobs/function_test/function_test.groovy")
-        }
 
+        def repo_dir = pwd()
+        def TESTS = "${env.TESTS}"
+        // Create an instance of UnitTest/UnitTest.groovy
+        def unit_test = load("jobs/UnitTest/UnitTest.groovy")
+        // Create an instance of FunctionTest/FunctionTest.groovy
+        def function_test = load("jobs/FunctionTest/FunctionTest.groovy")
+
+        try{
+            stage("Unit Test"){
+                // Call the function runTest to run unit test
+                unit_test.runTest(env.stash_manifest_name, env.stash_manifest_path, repo_dir)
+            }
+            stage("Function Test${i}"){
+                // Call the function run to run function test
+                function_test.run(TESTS, env.stash_manifest_name, env.stash_manifest_path, repo_dir)
+            }
+        } finally{
+            unit_test.archiveArtifactsToTarget("UnitTest")
+            function_test.archiveArtifactsToTarget("FunctionTest")
+        }
         shareMethod.buildAndPublish()
     } finally{
         shareMethod.sendResult(true, true)
